Move pprof to MetricsServer
