+++ cmd/cortex-gw/main.go
@@ -103,7 +103,6 @@ func main() {
 	initRoutes(api, writeProxy, *enforceRoles)
 
 	ms := util.NewMetricsServer(*metricsAddr)
-	util.StartDebugServer()
 
 	interrupt := make(chan os.Signal, 1)
 	signal.Notify(interrupt, os.Interrupt, syscall.SIGTERM)
